Use our custom `yaml_load_safely` for mkdocs.yml
