+++ readthedocs/config/utils.py
@@ -27,29 +27,3 @@ def list_to_dict(list_):
         for i, element in enumerate(list_)
     }
     return dict_
-
-
-class SafeLoaderIgnoreUnknown(yaml.SafeLoader):  # pylint: disable=too-many-ancestors
-
-    """
-    YAML loader to ignore unknown tags.
-
-    Borrowed from https://stackoverflow.com/a/57121993
-    """
-
-    def ignore_unknown(self, node):  # pylint: disable=no-self-use, unused-argument
-        return None
-
-
-SafeLoaderIgnoreUnknown.add_constructor(None, SafeLoaderIgnoreUnknown.ignore_unknown)
-
-
-def yaml_load_safely(content):
-    """
-    Uses ``SafeLoaderIgnoreUnknown`` loader to skip unknown tags.
-
-    When a YAML contains ``!!python/name:int`` it will complete ignore it an
-    return ``None`` for those fields instead of failing. We need this to avoid
-    executing random code, but still support these YAML files.
-    """
-    return yaml.load(content, Loader=SafeLoaderIgnoreUnknown)
