+++ readthedocs/doc_builder/backends/mkdocs.py
@@ -13,7 +13,6 @@ from django.conf import settings
 from django.template import loader as template_loader
 from readthedocs.projects.constants import MKDOCS_HTML, MKDOCS
 
-from readthedocs.config.utils import yaml_load_safely
 from readthedocs.doc_builder.base import BaseBuilder
 from readthedocs.doc_builder.exceptions import MkDocsYAMLParseError
 from readthedocs.projects.models import Feature
@@ -325,3 +324,29 @@ class MkdocsJSON(BaseMkdocs):
     type = 'mkdocs_json'
     builder = 'json'
     build_dir = '_build/json'
+
+
+class SafeLoaderIgnoreUnknown(yaml.SafeLoader):  # pylint: disable=too-many-ancestors
+
+    """
+    YAML loader to ignore unknown tags.
+
+    Borrowed from https://stackoverflow.com/a/57121993
+    """
+
+    def ignore_unknown(self, node):  # pylint: disable=no-self-use, unused-argument
+        return None
+
+
+SafeLoaderIgnoreUnknown.add_constructor(None, SafeLoaderIgnoreUnknown.ignore_unknown)
+
+
+def yaml_load_safely(content):
+    """
+    Uses ``SafeLoaderIgnoreUnknown`` loader to skip unknown tags.
+
+    When a YAML contains ``!!python/name:int`` it will complete ignore it an
+    return ``None`` for those fields instead of failing. We need this to avoid
+    executing random code, but still support these YAML files.
+    """
+    return yaml.load(content, Loader=SafeLoaderIgnoreUnknown)
